+++ scripts/restart.sh
@@ -3,7 +3,7 @@
 readonly TASK_NAME="Lux"
 
 # Kill the current task
-pkill -f "$TASK_NAME"
+if pgrep "$TASK_NAME"; then pkill -f "$TASK_NAME"; fi
 
 # Start it up again
 bash -c "exec -a $TASK_NAME ./start.sh"
